% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{euc_dist}
\alias{euc_dist}
\title{Euclidean distance calculation}
\usage{
euc_dist(x1, x2)
}
\arguments{
\item{x1}{A numeric vector of locations}

\item{x2}{A reference location with the same number of dimensions as x1}
}
\value{
Will return a numeric vector for a single reference location
}
\description{
Euclidean distance calculation
}
\examples{
x1 <- matrix(c(0,0,0,1,1,1,2,2,2), nrow = 3, byrow = TRUE)
x2 <- c(1, 1, 1)
euc_dist(x1, x2)
}
\keyword{internal}
