% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_geodist3d.R
\name{geo_dist_3d_calc}
\alias{geo_dist_3d_calc}
\title{Euclidean distance calculation}
\usage{
geo_dist_3d_calc(x, a, b, c)
}
\arguments{
\item{x}{A numeric vector of locations}

\item{a}{The reference longitude.}

\item{b}{The reference latitude.}

\item{c}{The reference height.}
}
\value{
Will return a numeric vector for a single reference location, or a
  matrix for multiple locations with (n_locations, n_distances).
}
\description{
Euclidean distance calculation
}
\keyword{internal}
