% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{norm}
\alias{norm}
\title{Parameter functions for spatial feature engineering recipes}
\usage{
norm(values = values_norm_func)
}
\arguments{
\item{values}{A character string of possible values. Either "l1", "l2", or
"max".}
}
\value{
A function with classes "qual_param" and "param"
}
\description{
New features derived from spatial distances require normalization. `norm` is
for specifying the type of row-normalization.
}
\examples{
norm("l1")
}
