% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_spatial_neighbors.R
\name{return_neighbours}
\alias{return_neighbours}
\title{Compute which points in `y` are the k-neighbors of `x`}
\usage{
return_neighbours(formula, x, y, k)
}
\arguments{
\item{formula}{A formula.}

\item{x}{A data.frame of training data.}

\item{y}{A data.frame of data to predict, can be the same as `x`.}

\item{k}{An integer specifying the number of neighbors to use.}
}
\value{

}
\description{
If `x` and `y` represent identical datasets then the points where x[i] ==
y[i] are not considered neighbors, i.e. the query point is not considered to
be its own neighbor.
}
