% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_clusterdist.R
\name{step_clusterdist}
\alias{step_clusterdist}
\alias{tidy.step_clusterdist}
\title{Distance between locations in two dimensions}
\usage{
step_clusterdist(
  recipe,
  ...,
  ref_lon,
  ref_lat,
  num_comp = 5,
  role = "predictor",
  trained = FALSE,
  name = "clusterdist",
  columns = NULL,
  centers = NULL,
  skip = FALSE,
  id = recipes::rand_id("clusterdist")
)

\method{tidy}{step_clusterdist}(x, ...)
}
\arguments{
\item{...}{One or more selector functions to choose which variables are
affected by the step. See selections() for more details. For the tidy
method, these are not currently used.}

\item{ref_lon}{A character with the name of the column in the training data
that represents that x-coordinate of the spatial domain.}

\item{ref_lat}{A character with the name of the column in the training data
that represents that y-coordinate of the spatial domain.}

\item{num_comp}{Number of clusters to generate.}

\item{role}{or model term created by this step, what analysis role should be
assigned?. By default, the function assumes that resulting distance will be
used as a predictor in a model.}

\item{name}{A single character value to use for the new predictor column. If
a column exists with this name, an error is issued. If multiple clusters
are used, then the name is appended with `1, 2, 3 ...` representing the
index of each reference location.}

\item{columns}{A character string of variable names that will be populated
(eventually) by the `terms` argument.}

\item{centers}{The cluster geographic locations of the cluster centroids
based on `ref_lon,` `ref_lat` and the clusters generated from the columns
specified by `terms`.}

\item{x}{A `step_clusterdist` object.}
}
\value{
An updated version of `recipe` with the new step added to the
  sequence of existing steps (if any).
}
\description{
`step_clusterdist` creates a *specification* of a recipe step that will
calculate the distance between points on a map to a reference location.
}
\concept{preprocessing}
\keyword{datagen}
