% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_geodist3d.R
\name{step_geodist3d}
\alias{step_geodist3d}
\alias{tidy.step_geodist3d}
\title{Distance between two locations}
\usage{
step_geodist3d(
  recipe,
  lat = NULL,
  lon = NULL,
  z = NULL,
  ref_lat = NULL,
  ref_lon = NULL,
  ref_z = NULL,
  role = "predictor",
  trained = FALSE,
  log = FALSE,
  name = "geo_dist3d",
  columns = NULL,
  skip = FALSE,
  id = recipes::rand_id("geodist3d")
)

\method{tidy}{step_geodist3d}(x, ...)
}
\arguments{
\item{lon, lat, z}{Selector functions to choose which variables are
affected by the step. See selections() for more details.}

\item{ref_lon, ref_lat, ref_z}{Single numeric values for the location
of the reference point.}

\item{role}{or model term created by this step, what analysis
role should be assigned?. By default, the function assumes
that resulting distance will be used as a predictor in a model.}

\item{log}{A logical: should the distance be transformed by
the natural log function?}

\item{name}{A single character value to use for the new
predictor column. If a column exists with this name, an error is
issued.}

\item{columns}{A character string of variable names that will
be populated (eventually) by the `terms` argument.}

\item{x}{A `step_geodist3d` object.}
}
\value{
An updated version of `recipe` with the new step added
 to the sequence of existing steps (if any). For the `tidy`
 method, a tibble with columns echoing the values of `lat`,
 `lon`, `z`, `ref_lat`, `ref_lon`, `ref_z`, `name`, and `id`.
}
\description{
`step_geodist3d` creates a *specification* of a
 recipe step that will calculate the distance between
 points on a map to a reference location.
}
\concept{preprocessing}
\keyword{datagen}
