% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_knn.R
\name{step_knn}
\alias{step_knn}
\alias{tidy.step_knn}
\title{Spatial lag step}
\usage{
step_knn(
  recipe,
  ...,
  outcome = NULL,
  role = "predictor",
  trained = FALSE,
  neighbors = 3,
  weight_func = "rectangular",
  dist_power = 2,
  scale = TRUE,
  data = NULL,
  columns = NULL,
  means = NULL,
  sds = NULL,
  skip = FALSE,
  id = recipes::rand_id("knn")
)

\method{tidy}{step_knn}(x, ...)
}
\arguments{
\item{recipe}{A recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See selections() for more details. For the tidy
method, these are not currently used.}

\item{outcome}{Selector function to choose which variable will be used to
create a new feature based on the inverse distance-weighted mean of
surrounding observations.}

\item{role}{role or model term created by this step, what analysis role
should be assigned?. By default, the function assumes that resulting
distance will be used as a predictor in a model.}

\item{trained}{A logical that will be updated once the step has been trained.}

\item{neighbors}{The number of closest neighbours to use in the distance
weighting.}

\item{weight_func}{A single character for the kernel function used to weight
the distances between samples. The default is 'rectangular' and the available
choices are 'rectangular', 'inv', 'gaussian'.}

\item{dist_power}{Power function for "inv". The default is 2.}

\item{scale}{A boolean to indicate whether the selected columns should be standardized
before transformation. Default is `TRUE`.}

\item{data}{Used internally to store the training data.}

\item{columns}{A character string that contains the names of columns used in the
transformation. This is `NULL` until computed by `prep.recipe()`.}

\item{means}{A named numeric vector of means. This is `NULL` until computed by
`prep.recipe()`.}

\item{sds}{A named numeric vector of standard deviations. This is `NULL` until
computed by `prep.recipe()`.}

\item{skip}{A logical to skip training.}

\item{id}{An identifier for the step. If omitted then this is generated
automatically.}

\item{x}{A `step_knn` object.}
}
\value{
An updated version of `recipe` with the new step added to the
  sequence of existing steps (if any).
}
\description{
`step_knn` creates a *specification* of a recipe step that will add a
new 'lag' feature to a dataset based on the weighted or unweighted mean or mode of
neighbouring observations.
}
\details{
No details yet!
}
\examples{
library(modeldata)
library(recipes)
library(spatialrecipes)
data(ames)

rec_obj <- ames \%>\%
recipe(Sale_Price ~ Latitude + Longitude) \%>\%
step_knn(Latitude, Longitude, outcome = "Sale_Price", neighbors = 3,
                 weight_func = "inv", dist_power = 2)

prepped <- prep(rec_obj)
juice(prepped)
}
\concept{preprocessing}
\keyword{datagen}
