% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_neighbors.R
\name{step_neighbors}
\alias{step_neighbors}
\title{Spatial si step}
\usage{
step_neighbors(
  recipe,
  ...,
  outcome = NULL,
  role = "predictor",
  trained = FALSE,
  neighbors = 3,
  data = NULL,
  columns = NULL,
  skip = FALSE,
  id = recipes::rand_id("neighbors")
)
}
\arguments{
\item{recipe}{A recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See selections() for more details. For the tidy
method, these are not currently used.}

\item{outcome}{Selector function to choose which variable will be used to
create a new feature based on the inverse distance-weighted mean of
surrounding observations.}

\item{role}{role or model term created by this step, what analysis
role should be assigned?. By default, the function assumes
that resulting distance will be used as a predictor in a model.}

\item{trained}{A logical that will be updated once the step has been trained.}

\item{neighbors}{The number of closest neighbors to use in the distance
weighting. The default is 3.}

\item{data}{Used internally to store the training data.}

\item{columns}{A character string that contains the names of columns used in the
transformation. This is `NULL` until computed by `prep.recipe()`.}

\item{skip}{A logical to skip training.}

\item{id}{An identifier for the step. If omitted then this is generated
automatically.}
}
\value{
An updated version of `recipe` with the new step added to the
  sequence of existing steps (if any).
}
\description{
`step_neighbors` creates a *specification* of a recipe step that will add a new
'si' features to a dataset based on the inverse distance-weighted mean of
surrounding observations.
}
