% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_spatial_clusterdist.R
\name{step_spatial_clusterdist}
\alias{step_spatial_clusterdist}
\alias{tidy.step_spatial_clusterdist}
\title{Distance between locations in two dimensions}
\usage{
step_spatial_clusterdist(
  recipe,
  ...,
  ref,
  num_comp = 5,
  role = "predictor",
  trained = FALSE,
  name = "spatial_clusterdist",
  columns = NULL,
  centers = NULL,
  skip = FALSE,
  id = recipes::rand_id("spatial_clusterdist")
)

\method{tidy}{step_spatial_clusterdist}(x, ...)
}
\arguments{
\item{...}{One or more selector functions to choose which variables are
affected by the step. See selections() for more details. For the tidy
method, these are not currently used.}

\item{ref}{A character with the name(s) of the columns in the training data
that represents the spatial domain. For example,
`ref = c("lat", "lon", "depth")` would calculate the centroids of clusters
using the terms specified in the selector function, and then would measure
the euclidean distances to these cluster centroids based on the `ref`
variables. The variables specified in `ref` also need to be included
as terms in the selector function.}

\item{num_comp}{Number of clusters to generate.}

\item{role}{or model term created by this step, what analysis role should be
assigned?. By default, the function assumes that resulting distance will be
used as a predictor in a model.}

\item{name}{A single character value to use for the new predictor column. If
a column exists with this name, an error is issued. If multiple clusters
are used, then the name is appended with `1, 2, 3 ...` representing the
index of each reference location.}

\item{columns}{A character string of variable names that will be populated
(eventually) by the `terms` argument.}

\item{centers}{The cluster geographic locations of the cluster centroids
based on `ref,` and the clusters generated from the columns
specified by `terms`.}

\item{x}{A `step_spatial_clusterdist` object.}
}
\value{
An updated version of `recipe` with the new step added to the
  sequence of existing steps (if any).
}
\description{
`step_spatial_clusterdist` creates a *specification* of a recipe step that will
calculate the distance between points on a map to a reference location.
}
\examples{
library(tidyverse)
library(tidymodels)

rec_obj <- Sacramento \%>\%
recipe(price ~ .) \%>\%
  step_spatial_clusterdist(
    price, latitude, longitude,
    ref = c("longitude", "latitude"),
    num_comp = 5
 )

 prepped <- prep(rec_obj
 bake(prepped, new_data = NULL)
}
\concept{preprocessing}
\keyword{datagen}
