% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_spatial_dist3d.R
\name{step_spatial_dist3d}
\alias{step_spatial_dist3d}
\alias{tidy.step_spatial_dist3d}
\title{Distance between locations in three dimensions}
\usage{
step_spatial_dist3d(
  recipe,
  lat = NULL,
  lon = NULL,
  height = NULL,
  ref_lat = NULL,
  ref_lon = NULL,
  ref_height = NULL,
  minimum = FALSE,
  log = FALSE,
  role = "predictor",
  trained = FALSE,
  name = "geo_dist3d",
  columns = NULL,
  skip = FALSE,
  id = recipes::rand_id("geodist3d")
)

\method{tidy}{step_spatial_dist3d}(x, ...)
}
\arguments{
\item{lon, lat, height}{Selector functions to choose which variables are
affected by the step. See selections() for more details.}

\item{ref_lon, ref_lat, ref_height}{Numeric values for the location of the
reference points. New features representing the euclidean distance to each
`ref_lon`, `ref_lat` and `ref_height` location will be created, unless
`minimum = TRUE` when the minimum distance to the combined features will be
created as a single new feature.}

\item{minimum}{A logical for whether only the minimum distance to the
reference locations should be calculated.}

\item{log}{A logical: should the distance be transformed by the natural log
function?}

\item{role}{or model term created by this step, what analysis role should be
assigned?. By default, the function assumes that resulting distance will be
used as a predictor in a model.}

\item{name}{A single character value to use for the new predictor column. If
a column exists with this name, an error is issued. If multiple reference
locations are used, then the name is appended with `1, 2, 3 ...`
representing the index of each reference location.}

\item{columns}{A character string of variable names that will be populated
(eventually) by the `terms` argument.}

\item{x}{A `step_spatial_dist3d` object.}
}
\value{
An updated version of `recipe` with the new step added to the
  sequence of existing steps (if any). For the `tidy` method, a tibble with
  columns echoing the values of `lat`, `lon`, `height`, `ref_lat`, `ref_lon`,
  `ref_height`, `name`, and `id`.
}
\description{
`step_spatial_dist3d` creates a *specification* of a recipe step that will
calculate the distance between points on a map to a reference location.
}
\concept{preprocessing}
\keyword{datagen}
