% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_spatial_lag.R
\name{step_spatial_lag}
\alias{step_spatial_lag}
\alias{tidy.step_spatial_lag}
\title{Spatial lag step}
\usage{
step_spatial_lag(
  recipe,
  ...,
  outcome = NULL,
  role = "predictor",
  trained = FALSE,
  neighbors = NA,
  weight_func = NULL,
  dist = FALSE,
  data = NULL,
  columns = NULL,
  skip = FALSE,
  id = recipes::rand_id("spatial_lag")
)

\method{tidy}{step_spatial_lag}(x, ...)
}
\arguments{
\item{recipe}{A recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See selections() for more details. For the tidy
method, these are not currently used.}

\item{outcome}{Selector function to choose which variable will be used to
create a new feature based on the inverse distance-weighted mean of
surrounding observations.}

\item{role}{role or model term created by this step, what analysis role
should be assigned?. By default, the function assumes that resulting
distance will be used as a predictor in a model.}

\item{trained}{A logical that will be updated once the step has been trained.}

\item{neighbors}{The number of closest neighbours to use in the distance
weighting.}

\item{weight_func}{A single character for the kernel function used to weight
the distances between samples. Valid choices are: "rectangular",
"triangular", "epanechnikov", "biweight", "tri-weight", "cos", "inv",
"gaussian", "rank", and "optimal".}

\item{dist}{Whether to also return the weighted mean of the distances to the
neighbours.}

\item{data}{Used internally to store the training data.}

\item{skip}{A logical to skip training.}

\item{id}{An identifier for the step. If omitted then this is generated
automatically.}

\item{x}{A `step_spatial_lag` object.}
}
\value{
An updated version of `recipe` with the new step added to the
  sequence of existing steps (if any).
}
\description{
`step_spatial_lag` creates a *specification* of a recipe step that will add a
new 'lag' feature to a dataset based on the inverse distance-weighted mean of
surrounding observations.
}
\details{
No details yet!
}
\examples{
No examples yet!
}
\concept{preprocessing}
\keyword{datagen}
